Refactor db.rs: Extract tests and restoration logic
- Extracted ~860 lines of tests to `src/db/tests.rs`
- Extracted ~220 lines of index restoration logic to `src/db/restoration.rs`
- Cleaned up `GallifreyDB::with_unified_config`
- Verified with existing tests
